backend/requirements/dev.txt
flask
flask-sqlalchemy
flask-marshmallow
sqlalchemy
marshmallow_sqlalchemy
python-dotenv
marshmallow
marshmallow-enum
flask-jwt-extended
werkzeug
flask-restful
flask-bcrypt
flask-cors
celery[redis]
Flask-PyMongo